Modificacion de metodo para editar productos

+++ src/controllers/product.controller.js
@@ -94,14 +94,13 @@ const productController = {
     
         try {
             const product = await productService.getProductDetail(productId);
-            const user = await userService.getUserById(userId)
-
+            const user = await userService.getUserById(userId);
+    
             if (userRole === 'admin' || (userRole === 'premium' && user && user._id.toString() === product.owner._id.toString())) {
                 const updatedProduct = await productService.updateProduct(productId, req, productUpdateData, userId);
-
+    
                 return res.json({ message: "Producto actualizado!", product: updatedProduct });
-            }
-            else{
+            } else {
                 return res.status(403).json({ message: 'No tienes permiso para realizar esta acción' });
             }
         } catch (err) {
@@ -132,24 +131,22 @@ const productController = {
         const productId = req.params.pid;
         const userId = req.session.userId;
         const userRole = req.session.userRole;
-
+    
         try {
             const product = await productService.getProductDetail(productId);
-            const user = await userService.getUserById(userId)
-
-            if (userRole == 'admin' || (userRole == 'premium' && user == product.owner)) {
+            const user = await userService.getUserById(userId);
+    
+            if (userRole === 'admin' || (userRole === 'premium' && user && user._id.toString() === product.owner._id.toString())) {
                 await productService.deleteProduct(productId);
-
                 return res.json({ message: "Producto eliminado!" });
-            }
-            else{
+            } else {
                 return res.status(403).json({ message: 'No tienes permiso para realizar esta acción' });
             }
         } catch (err) {
             console.error('Error:', err);
             return res.status(500).json({ error: "Error en la base de datos", details: err.message });
         }
-    }
+    }  
 }
 
 export default productController;
